+++ src/Container.php
@@ -106,10 +106,10 @@ class Container implements ContainerInterface, FactoryInterface, InvokerInterfac
 
         // Auto-register the container
         $this->resolvedEntries = [
-            self::class => $this,
+            self::class => $this->delegateContainer,
+            ContainerInterface::class => $this->delegateContainer,
             FactoryInterface::class => $this,
             InvokerInterface::class => $this,
-            ContainerInterface::class => $this,
         ];
     }